requirements.txt
Django==2.2.4
djangorestframework==3.10.2
djangorestframework-stubs==1.0.0
psycopg2==2.7.4
pytz==2019.2
sqlparse==0.3.0
